Move functions to separate file for easier unit testing
